radius_proxy: fix buffer overflows on large packets (#6)
Summary:
Pull Request resolved: https://github.com/facebookexternal/fbc_shared/pull/6

If the size of a datagram uses all of BUFLEN (4096 bytes) then there
are a couple of places where memcpy will overrun a buffer.  Fix these
by adding missing bounds checks.

In these cases we are just copying the packets in order to log them
so truncating to radius_packet_t size poses no additional issues.

Found with afl-fuzz.

Differential Revision: D14084675

fbshipit-source-id: c5029362624d0dee3bdaec39e31af941df13160c